tracking-manager/packages/payment/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
qrcode==7.4.2
pillow==10.1.0
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.20.0
pymongo~=4.9.2